Moved from Vector3d to Vector3i

+++ src/main/java/net/felixoi/felograms/internal/configuration/serializer/HologramTypeSerializer.java
@@ -1,6 +1,7 @@
 package net.felixoi.felograms.internal.configuration.serializer;
 
 import com.flowpowered.math.vector.Vector3d;
+import com.flowpowered.math.vector.Vector3i;
 import com.google.common.reflect.TypeToken;
 import net.felixoi.felograms.Felograms;
 import net.felixoi.felograms.api.hologram.Hologram;
@@ -22,7 +23,7 @@ public class HologramTypeSerializer implements TypeSerializer<Hologram> {
         List<Text> lines = value.getNode("lines").getList(new TypeToken<Text>() {
         });
         UUID worldUUID = value.getNode("world").getValue(TypeToken.of(UUID.class));
-        Vector3d position = value.getNode("position").getValue(TypeToken.of(Vector3d.class));
+        Vector3i position = value.getNode("position").getValue(TypeToken.of(Vector3i.class));
         boolean disabled = value.getNode("disabled").getBoolean();
 
         return new SimpleHologram(Felograms.getInstance().getHologramManager(), uuid, name, lines, worldUUID, position, disabled);
